miscellaneous debug changes
