+++ Inc/ST-LIB_LOW/Sensors/AnalogSensor/AnalogSensor.hpp
@@ -19,4 +19,5 @@ public:
 protected:
 	Pin pin;
 	uint8_t id;
+	double *value;
 };
