+++ Inc/ST-LIB_LOW/Sensors/LinearSensor/LinearSensor.hpp
@@ -22,5 +22,4 @@ protected:
 	double slope;
 	double offset;
 	double *value;
-
 };
