+++ Src/ST-LIB_LOW/Sensors/LinearSensor/LinearSensor.cpp
@@ -8,6 +8,7 @@ LinearSensor::LinearSensor(Pin pin, double slope, double offset, double *value)
 		id = identification.value();
 		ADC::turn_on_adc(id);
 	}else{
+		id = 0xFFFF;
 		//errores de optional vacio aqui (por hacer)
 	}
 }
@@ -17,6 +18,7 @@ void LinearSensor::read(){
 	if(val){
 		*value = slope * (double) val.value() + offset;
 	}else{
+		*value = INFINITY;
 		//errores de optional vacio aqui (por hacer)
 	}
 }
